+++ test/config/database.js
@@ -1,7 +1,7 @@
 /**
  * Database configuration.
  */
-export default {
+ module.exports = {
   development: {
     username: 'root',
     password: null,