+++ types/interfaces/FaceMatch.d.ts
@@ -22,4 +22,10 @@ export default interface  {
      * @type {string}
      */
     externalImageId?: string;
+    /**
+     * Confidence in the match of this face with the input face.
+     * Valid Range: Minimum value of 0. Maximum value of 100.
+     * @type {number}
+    */
+    similarity?: number;
 }