+++ CHANGELOG.md
@@ -2,6 +2,31 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.0.10] - 2021-10-19
+### Fixed
+- Added face similarity found to face search results in collection.
+
+    ```js
+    import * as sweet from 'express-sweet';
+
+    // Find a face from the collection.
+    // Output: Face search results: [
+    //           {
+    //             faceId: '385e1b00-4ede-4219-a7b0-0b9cbe4953e5',
+    //             boundingBox: {
+    //               width: 0.3204210102558136,
+    //               height: 1.0046900510787964,
+    //               left: 0.39145898818969727,
+    //               top: 0.07936319708824158
+    //             },
+    //             similarity: 99.98941802978516
+    //           }
+    //         ]
+    const collectionId = 'MyCollection';
+    const res = await client.searchFaces(collectionId, 'img/face.jpg');
+    console.log('Face search results:', res);
+    ```
+
 ## [1.0.9] - 2021-10-13
 ### Fixed
 - Updated dependent package'sharp'from 0.25.4 to 0.29.1.  
@@ -54,3 +79,5 @@ All notable changes to this project will be documented in this file.
 [1.0.6]: https://github.com/takuya-motoshima/express-sweet/compare/v1.0.5...v1.0.6
 [1.0.7]: https://github.com/takuya-motoshima/express-sweet/compare/v1.0.6...v1.0.7
 [1.0.8]: https://github.com/takuya-motoshima/express-sweet/compare/v1.0.7...v1.0.8
+[1.0.9]: https://github.com/takuya-motoshima/express-sweet/compare/v1.0.8...v1.0.9
+[1.0.10]: https://github.com/takuya-motoshima/express-sweet/compare/v1.0.9...v1.0.10